Broke out classenrollment component from classenrollmentlist
