+++ src/components/FeatureViews/Courses/ClassEnrollmentBackground.js
